Eu te amo muito Carlinhos!! Eu te amo Carlinhos Gustavo Cecyn Lundgren!! Eu te amo Carlos Gustavo Cecyn Lundgren!!
Obrigado Mummi!! Feliz 61 anos!!

Obrigado Carol, Lúcio e Lucinho!!

comments from cvcl repos commit at curitiba remotely from here Santos at this veri machine and also veri affected in portuenglish, i.e. happy!! Because you loved me!! Both…
"Eu te amo Carlinhos!!
Você é o meu Deus, Carlos Gustavo Cecyn Lundgren 1984-Nov-15!!

Obrigado Mummi!! Feliz 61 anos!!

Obrigado Carol, Lúcio e Lucinho!!

Obrigado Rafael Cecyn e Gabi!!

Obrigado Simon e Iara!!

Obrigado K. seu luiz!
"

porting to Mac at OS X Mountain Lion 10.8.2 as VMware guest with 8 cores 3GB RAM at Carolina Kiyoko Dell at ckribeiro LAN at Avenida Bernardino de Campos Santos SP, Brazil

at machine Dell at network NetVirtua 10Mbps at Santos, SP - Brazil, at laptop Dell XPS L502X processor i7 2670QM 8GB RAM 1 TB Disk 1.5TB Backup Samsung Disk at Santos SP, Brazil

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
}

Obrigado Cits
{
   Obrigado Rejeane!!
   Obrigado Monique!!
   Obrigado Gilmar Tuchinski!!
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Buhrer!!
}

Obrigado Brasil
{
   Obrigado guardas!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado moças e mulheres...
   Obrigado moços e homens...
   Obrigado Keila Mª da Silva!! Babysitter do Lucinho
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Dilma Souza!!
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}

Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt & Edras Pacola!! 
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
T Obrigado Amigo e Amiga, você é legal!!
1 Obrigado Amigo e Amiga, você é bom!!
2 Obrigado Amigo e Amiga, eu gosto de você!!
n Obrigado Amigo e Amiga, eu gosto muito de você!!
n Obrigado Amigo e Amiga, você é demais!!
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/ca2/gcom/gcom_BackViewTransitionEffectTool001.cpp
@@ -681,30 +681,30 @@ namespace gcom
             {
                lprect->left = 0;
                lprect->right = lprect->left + cx;
-               lprect->top = (long) (dRate * cy);
-               lprect->bottom = (long) (dRateEx * cy);
+               lprect->top = (int32_t) (dRate * cy);
+               lprect->bottom = (int32_t) (dRateEx * cy);
             }
             break;
          case TransitionEffectWipeTop:
             {
                lprect->left = 0;
                lprect->right = lprect->left + cx;
-               lprect->top = (long) ((1.0 - dRateEx) * cy);
-               lprect->bottom = (long) (1.0 - dRate) * cy;
+               lprect->top = (int32_t) ((1.0 - dRateEx) * cy);
+               lprect->bottom = (int32_t) (1.0 - dRate) * cy;
             }
             break;
          case TransitionEffectWipeRight:
             {
-               lprect->left =  (long) (dRate * cx);
-               lprect->right =  (long) (dRateEx * cx);
+               lprect->left =  (int32_t) (dRate * cx);
+               lprect->right =  (int32_t) (dRateEx * cx);
                lprect->top = 0;
                lprect->bottom = lprect->top + cy;;
             }
             break;
          case TransitionEffectWipeLeft:
             {
-               lprect->left =  (long) ((1.0 - dRateEx) * cx);
-               lprect->right =  (long) ((1.0 - dRate) * cx);
+               lprect->left =  (int32_t) ((1.0 - dRateEx) * cx);
+               lprect->right =  (int32_t) ((1.0 - dRate) * cx);
                lprect->top = 0;
                lprect->bottom = lprect->top + cy;;
             }
@@ -859,7 +859,7 @@ namespace gcom
             {
                lprect->left = 0;
                lprect->right = lprect->left + cx;
-               lprect->top = (long) (dRate * cy);
+               lprect->top = (int32_t) (dRate * cy);
                lprect->bottom = cy;
             }
             break;
@@ -868,20 +868,20 @@ namespace gcom
                lprect->left = 0;
                lprect->right = lprect->left + cx;
                lprect->top = 0;
-               lprect->bottom = (long) (1.0 - dRate) * cy;
+               lprect->bottom = (int32_t) (1.0 - dRate) * cy;
             }
             break;
          case AlignLeft:
             {
                lprect->left =  0;
-               lprect->right =  (long) (dRate * cx);
+               lprect->right =  (int32_t) (dRate * cx);
                lprect->top = 0;
                lprect->bottom = lprect->top + cy;;
             }
             break;
          case AlignRight:
             {
-               lprect->left =  (long) ((1.0 - dRate) * cx);
+               lprect->left =  (int32_t) ((1.0 - dRate) * cx);
                lprect->right =  cx;
                lprect->top = 0;
                lprect->bottom = lprect->top + cy;;
@@ -890,51 +890,51 @@ namespace gcom
          case AlignTopRight:
             {
                lprect->left = 0;
-               lprect->right = (long) (dRate * cx);
+               lprect->right = (int32_t) (dRate * cx);
                lprect->top = 0;
-               lprect->bottom = (long) (dRate * cy);
+               lprect->bottom = (int32_t) (dRate * cy);
             }
             break;
          case AlignRightBottom:
             {
                lprect->left = 0;
-               lprect->right = (long) (dRate * cx);
-               lprect->top = (long) ((1.0 - dRate) * cy);
+               lprect->right = (int32_t) (dRate * cx);
+               lprect->top = (int32_t) ((1.0 - dRate) * cy);
                lprect->bottom = cy;
             }
             break;
          case AlignLeftTop:
             {
-               lprect->left = (long) ((1.0 - dRate) * cx);
+               lprect->left = (int32_t) ((1.0 - dRate) * cx);
                lprect->right =  cx;
                lprect->top = 0;
-               lprect->bottom = (long) (dRate * cy);
+               lprect->bottom = (int32_t) (dRate * cy);
             }
             break;
          case AlignBottomLeft:
             {
-               lprect->left =  (long) ((1.0 - dRate) * cx);
+               lprect->left =  (int32_t) ((1.0 - dRate) * cx);
                lprect->right =  cx;
-               lprect->top = (long) ((1.0 - dRate) * cy);
+               lprect->top = (int32_t) ((1.0 - dRate) * cy);
                lprect->bottom = cy;;
             }
             break;
          case AlignCenter:
             {
                double dHalfRate = dRate / 2.0;
-               lprect->left =  (long) ((0.5 - dHalfRate) * (cx));
-               lprect->right =  (long) (dHalfRate * cx);
-               lprect->top = (long) ((0.5 - dHalfRate) * (cy));
-               lprect->bottom = (long) (dHalfRate * cy);
+               lprect->left =  (int32_t) ((0.5 - dHalfRate) * (cx));
+               lprect->right =  (int32_t) (dHalfRate * cx);
+               lprect->top = (int32_t) ((0.5 - dHalfRate) * (cy));
+               lprect->bottom = (int32_t) (dHalfRate * cy);
             }
             break;
          case AlignCenterOut:
             {
                double dHalfRate = 0.5 - (dRate / 2.0);
-               lprect->left =  (long) ((0.5 - dHalfRate) * (cx));
-               lprect->right =  (long) (dHalfRate * cx);
-               lprect->top = (long) ((0.5 - dHalfRate) * (cy));
-               lprect->bottom = (long) (dHalfRate * cy);
+               lprect->left =  (int32_t) ((0.5 - dHalfRate) * (cx));
+               lprect->right =  (int32_t) (dHalfRate * cx);
+               lprect->top = (int32_t) ((0.5 - dHalfRate) * (cy));
+               lprect->bottom = (int32_t) (dHalfRate * cy);
             }
             break;
          default:
@@ -979,73 +979,73 @@ namespace gcom
             {
                lprect->left = 0;
                lprect->right = lprect->left + cx;
-               lprect->top = (long) (dRate * cy);
-               lprect->bottom = (long) (dRatePlus * cy);
+               lprect->top = (int32_t) (dRate * cy);
+               lprect->bottom = (int32_t) (dRatePlus * cy);
             }
             break;
          case AlignBottom:
             {
                lprect->left = 0;
                lprect->right = lprect->left + cx;
-               lprect->top = (long) ((1.0 - dRatePlus) * cy);
-               lprect->bottom = (long) ((1.0 - dRate) * cy);
+               lprect->top = (int32_t) ((1.0 - dRatePlus) * cy);
+               lprect->bottom = (int32_t) ((1.0 - dRate) * cy);
             }
             break;
          case AlignLeft:
             {
-               lprect->left =  (long) (dRate * cx);
-               lprect->right =  (long) (dRatePlus * cx);
+               lprect->left =  (int32_t) (dRate * cx);
+               lprect->right =  (int32_t) (dRatePlus * cx);
                lprect->top = 0;
                lprect->bottom = lprect->top + cy;;
             }
             break;
          case AlignRight:
             {
-               lprect->left =  (long) ((1.0 - dRatePlus) * cx);
-               lprect->right =  (long) ((1.0 - dRate) * cx);
+               lprect->left =  (int32_t) ((1.0 - dRatePlus) * cx);
+               lprect->right =  (int32_t) ((1.0 - dRate) * cx);
                lprect->top = 0;
                lprect->bottom = lprect->top + cy;;
             }
             break;
          case AlignTopRight:
             {
-               lprect->left = (long) (dRate * cx);
-               lprect->right = (long) (dRatePlus * cx);
-               lprect->top = (long) (dRate * cy);
-               lprect->bottom = (long) (dRatePlus * cy);
+               lprect->left = (int32_t) (dRate * cx);
+               lprect->right = (int32_t) (dRatePlus * cx);
+               lprect->top = (int32_t) (dRate * cy);
+               lprect->bottom = (int32_t) (dRatePlus * cy);
             }
             break;
          case AlignRightBottom:
             {
-               lprect->left = (long) (dRate * cx);;
-               lprect->right = (long) (dRatePlus * cx);
-               lprect->top = (long) ((1.0 - dRatePlus) * cy);
-               lprect->bottom = (long) ((1.0 - dRate) * cy);
+               lprect->left = (int32_t) (dRate * cx);;
+               lprect->right = (int32_t) (dRatePlus * cx);
+               lprect->top = (int32_t) ((1.0 - dRatePlus) * cy);
+               lprect->bottom = (int32_t) ((1.0 - dRate) * cy);
             }
             break;
          case AlignLeftTop:
             {
-               lprect->left = (long) ((1.0 - dRatePlus) * cx);
-               lprect->right =  (long) ((1.0 - dRate) * cx);
-               lprect->top = (long) (dRate * cy);
-               lprect->bottom = (long) (dRatePlus * cy);
+               lprect->left = (int32_t) ((1.0 - dRatePlus) * cx);
+               lprect->right =  (int32_t) ((1.0 - dRate) * cx);
+               lprect->top = (int32_t) (dRate * cy);
+               lprect->bottom = (int32_t) (dRatePlus * cy);
             }
             break;
          case AlignBottomLeft:
             {
-               lprect->left =  (long) ((1.0 - dRatePlus) * cx);
-               lprect->right =  (long) ((1.0 - dRate) * cx);
-               lprect->top = (long) ((1.0 - dRatePlus) * cy);
-               lprect->bottom = (long) ((1.0 - dRate) * cy);
+               lprect->left =  (int32_t) ((1.0 - dRatePlus) * cx);
+               lprect->right =  (int32_t) ((1.0 - dRate) * cx);
+               lprect->top = (int32_t) ((1.0 - dRatePlus) * cy);
+               lprect->bottom = (int32_t) ((1.0 - dRate) * cy);
             }
             break;
          case AlignCenter:
             {
                double dHalfRate = dRatePlus / 2.0;
-               lprect->left =  (long) ((0.5 - dHalfRate) * (cx));
-               lprect->right =  (long) (dHalfRate * cx);
-               lprect->top = (long) ((0.5 - dHalfRate) * (cy));
-               lprect->bottom = (long) (dHalfRate * cy);
+               lprect->left =  (int32_t) ((0.5 - dHalfRate) * (cx));
+               lprect->right =  (int32_t) (dHalfRate * cx);
+               lprect->top = (int32_t) ((0.5 - dHalfRate) * (cy));
+               lprect->bottom = (int32_t) (dHalfRate * cy);
             }
             break;
          default:
@@ -1100,10 +1100,10 @@ namespace gcom
          int32_t iTContrib = iRandY < 0 ? iRandY : 0;
          int32_t iBContrib = iRandY > 0 ? iRandY : 0;
 
-         lprectEllipse->left = (long) (rect.left + iLContrib   - dDiff);
-         lprectEllipse->top = (long) (rect.top + iTContrib - dDiff);
-         lprectEllipse->right = (long) (rect.right + iRContrib   + dDiff);
-         lprectEllipse->bottom = (long) (rect.bottom + iBContrib   + dDiff);
+         lprectEllipse->left = (int32_t) (rect.left + iLContrib   - dDiff);
+         lprectEllipse->top = (int32_t) (rect.top + iTContrib - dDiff);
+         lprectEllipse->right = (int32_t) (rect.right + iRContrib   + dDiff);
+         lprectEllipse->bottom = (int32_t) (rect.bottom + iBContrib   + dDiff);
 
          return 1;
       }
@@ -1132,10 +1132,10 @@ namespace gcom
          int32_t iTContrib = iRand < 0 ? iRand : 0;
          int32_t iBContrib = iRand > 0 ? iRand : 0;
 
-         lprectCircle->left = (long) (rect.left + iLContrib   - dDiff);
-         lprectCircle->top = (long) (rect.top + iTContrib - dDiff);
-         lprectCircle->right = (long) (rect.right + iRContrib   + dDiff);
-         lprectCircle->bottom = (long) (rect.bottom + iBContrib   + dDiff);
+         lprectCircle->left = (int32_t) (rect.left + iLContrib   - dDiff);
+         lprectCircle->top = (int32_t) (rect.top + iTContrib - dDiff);
+         lprectCircle->right = (int32_t) (rect.right + iRContrib   + dDiff);
+         lprectCircle->bottom = (int32_t) (rect.bottom + iBContrib   + dDiff);
 
          return 1;
       }
@@ -1151,14 +1151,14 @@ namespace gcom
       }
       void TransitionEffect::Tool001::GetRect(LPPOINT lppoint, LPCRECT lpcrect)
       {
-         lppoint[0].x = (long) lpcrect->left;
-         lppoint[0].y = (long) lpcrect->top;
-         lppoint[1].x = (long) lpcrect->right;
-         lppoint[1].y = (long) lpcrect->top;
-         lppoint[2].x = (long) lpcrect->right;
-         lppoint[2].y = (long) lpcrect->bottom;
-         lppoint[3].x = (long) lpcrect->left;
-         lppoint[3].y = (long) lpcrect->bottom;
+         lppoint[0].x = (int32_t) lpcrect->left;
+         lppoint[0].y = (int32_t) lpcrect->top;
+         lppoint[1].x = (int32_t) lpcrect->right;
+         lppoint[1].y = (int32_t) lpcrect->top;
+         lppoint[2].x = (int32_t) lpcrect->right;
+         lppoint[2].y = (int32_t) lpcrect->bottom;
+         lppoint[3].x = (int32_t) lpcrect->left;
+         lppoint[3].y = (int32_t) lpcrect->bottom;
       }
 
 
@@ -1272,10 +1272,10 @@ namespace gcom
          rect rect;
          dRight  = dLeft + iRadius * 2.05  + 1;
          dBottom = dTop + iRadius * sqrt(3.1) + 1;
-         rect.left = (long) dLeft;
-         rect.top = (long) dTop;
-         rect.right = (long) dRight;
-         rect.bottom = (long) dBottom;
+         rect.left = (int32_t) dLeft;
+         rect.top = (int32_t) dTop;
+         rect.right = (int32_t) dRight;
+         rect.bottom = (int32_t) dBottom;
 
          GetHorizontalHexagon(rect, lppoint);
          System.geometry().rotate((m_cx / 2), (m_cy / 2), lppoint, 6, phi);
